\name{benchmarksWM.overview}
\alias{benchmarksWM.overview}
\docType{data}
\encoding{UTF-8}
\title{An Overview of the Working Memory Research Benchmark Dataset Categories as in Oberauer et al. (2018)}

\description{
  An overview of the subsections of Benchmarks in Working Memory research
  as listed in Oberauer et al. (2018). The subsections include a description
  of the categories of findings as well as information about the respective
  datasets that were archived by Oberauer (2019).
}
\details{
  \strong{A list of categories of Benchmark data in Working Memory research from Oberauer
  et al. (2018).}
  \describe{
  \item{\code{\link{benchmarkWM1.1}}}{Set size effects on accuracy.}
  \item{\code{\link{benchmarkWM1.2}}}{Set size effects on response times.}
  \item{\code{\link{benchmarkWM1.3}}}{Capacity limits in verbal Working Memory.}
  \item{\code{\link{benchmarkWM2.1}}}{Effects of filled retention intervals.}
  \item{\code{\link{benchmarkWM2.4}}}{Effects of presentation duration of memory items.}
  \item{\code{\link{benchmarkWM3}}}{Effects of serial position.}
  \item{\code{\link{benchmarkWM4}}}{Characteristics of errors.}
  \item{\code{\link{benchmarkWM5}}}{Effects of multiple tasks and multiple stimulus domains.}
  \item{\code{\link{benchmarkWM6}}}{Effects of irrelevant sounds during WM tasks.}
  \item{\code{\link{benchmarkWM7}}}{Effects of syllable-based word length.}
  \item{\code{\link{benchmarkWM8}}}{Effects of phonological similarity and of item-probe similarity in change detection tasks.}
  \item{\code{\link{benchmarkWM9}}}{Effects of temporal distinctiveness and of grouping of lists.}
  \item{\code{\link{benchmarkWM10}}}{Effects of prioritization in WM (e.g. retro-cues)}
  \item{\code{\link{benchmarkWM11}}}{Effects of knowledge such as chunking or the Hebb repitition effect.}
  \item{\code{\link{benchmarkWM12}}}{Individual differences in WM performance and capacity.}
  }
}
\source{
  Oberauer, K. (2019). BenchmarksWM. https://github.com/oberauer/BenchmarksWM
}
\references{
  Oberauer, K., Lewandowsky, S., Awh, E., Brown, G.D.A, Conway, A.,
  Cowan, N., Donkin, C., Farrell, S., Hitch, G.J., Hurlstone, M.J., Ma,
  W.J., Morey, C.C., Nee, D.E., Schweppe, J., Vergauwe, E., and Ward,
  G. (2018). Benchmarks for models of short-term and working
  memory. \emph{Psychological Bulletin, 144}, 885-958.
}
\examples{
  help("benchmarksWM.overview")
  ?benchmarkWM1.1
  help("benchmarkWM3")
  data(unsworth06a)
}
\keyword{datasets}
