\name{benchmarkWM1.2}
\alias{benchmarkWM1.2}
\encoding{UTF-8}
\docType{data}

\title{Working Memory Benchmarks: Set Size Effects on Response Times}

\description{
  Summary dataset relating to the subsection of the Working Memory
  Benchmarks datasets (Oberauer, 2019) associated
  with set size effects on response time measures. Overall, the data show a
  monotonic increase in response times with increasing set size.
  See section "Details" for
  links to the documentation files providing more precise descriptions
  of the single datasets included in this Benchmark.

}
\details{
  \strong{4 datasets with set size effects on response times.}
  \describe{
    \item{\code{\link{donkin12b}}}{Set size and serial position effects
    on response times for fast vs. slow presentation rates in the Sternberg
    task.}
    \item{\code{\link{gilchrist14}}}{Set size effects on response times
    in a change detection task.}
    \item{\code{\link{shepherdson18}}}{Effects of retro cues and set size on
    response times in recognition memory with data collected by Souza, Rerko, & Oberauer (2014).}
    \item{\code{\link{towse08}}}{Effects of list length on response times in a
    complex span task.}
  }
}
\source{
  Oberauer, K. (2019). BenchmarksWM. https://github.com/oberauer/BenchmarksWM

}
\references{

  Donkin, C., & Nosofsky, R. M. (2012b). The structure of short-term memory
  scanning: An investigation using response time distribution models.
  \emph{Psychonomic Bulletin & Review, 19}, 363-394.
  http://dx.doi.org/ 10.3758/s13423-012-0236-8

  Oberauer, K., Lewandowsky, S., Awh, E., Brown, G.D.A, Conway, A.,
  Cowan, N., Donkin, C., Farrell, S., Hitch, G.J., Hurlstone, M.J., Ma,
  W.J., Morey, C.C., Nee, D.E., Schweppe, J., Vergauwe, E., and Ward,
  G. (2018). Benchmarks for models of short-term and working
  memory. \emph{Psychological Bulletin, 144}, 885-958.

  Shepherdson, P., Oberauer, K., & Souza, A. S. (2018). Working Memory Load
  and the Retro-Cue Effect: A Diffusion Model Account. \emph{Journal of
  Experimental Psychology: Human Perception and Performance, 44},
  286-310. http://dx.doi.org/10.1037/xhp0000448

  Souza, A. S., Rerko, L., & Oberauer, K. (2014). Unloading and reloading
  working memory: Attending to one item frees capacity.
  \emph{Journal of Experimental Psychology: Human Perception and
  Performance, 40}, 1237-1256. http://dx.doi.org/10.1037/a0036331

  Sternberg, S. (1966). High-speed scanning in human memory. \emph{Science,
  153}, 652-654. http://dx.doi.org/10.1126/science.153.3736.652

  Towse, J. N., Cowan, N., Hitch, G. J., & Horton, N. J. (2008). The recall of
  information from working memory. Insights from behavioural and chronometric
  perspectives. \emph{Experimental Psychology, 55}, 371-383.
  http://dx.doi.org/10.1027/1618-3169.55.6.371

}
\examples{
data(donkin12b)
?donkin12b
}
\keyword{datasets}
