\name{benchmarkWM7}
\alias{benchmarkWM7}
\docType{data}
\encoding{UTF-8}
\title{Working Memory Benchmark 7: Effects of Syllable-Based Word Length}
\description{
  Subsection of the Working Memory (WM) Benchmarks datasets (Oberauer, 2019) associated
  with the effects of syllable word length. This documentation summarizes the
  subsections of Benchmark 7 in Oberauer et al. (2018) and contains data on the
  effects of syllable-based word length and its boundary condition with articulatory
  suppression.
  See Oberauer et al. (2018; 906f.) and Bhatarah et al. (2009) for a more precise
  description of the datasets and the Benchmark findings they represent.
}
\details{
\strong{5 datasets of the 5 Experiments on the Word Length Effect from Bhatarah et al. (2009).}
\describe{
  \item{\code{\link{bhatarah09a}}}{Experiment 1: Differences between pre- and postcued immediate serial recall and free recall in verbal WM.}
  \item{\code{\link{bhatarah09b}}}{Experiment 2: The effects of word length and presentation rate.}
  \item{\code{\link{bhatarah09c}}}{Experiment 3: The effects of word length and articulatory suppression.}
  \item{\code{\link{bhatarah09d}}}{Experiment 4: The Word Length Effect with short lists of words.}
  \item{\code{\link{bhatarah09e}}}{Experiment 5: The Word Length Effect with long lists of words.}
  }
}
\source{
  Oberauer, K. (2019). BenchmarksWM. https://github.com/oberauer/BenchmarksWM
}
\references{

  Bhatarah, P., Ward, G., Smith, J., & Hayes, L. (2009). Examining the relationship between
  free recall and immediate serial recall: Similar patterns of rehearsal and similar effects
  of word length, presentation rate, and articulatory suppression. \emph{Memory & Cognition,
  37}, 689-713. http://dx.doi.org/10.3758/MC.37.5.689

  Oberauer, K., Lewandowsky, S., Awh, E., Brown, G.D.A, Conway, A.,
  Cowan, N., Donkin, C., Farrell, S., Hitch, G.J., Hurlstone, M.J., Ma,
  W.J., Morey, C.C., Nee, D.E., Schweppe, J., Vergauwe, E., and Ward,
  G. (2018). Benchmarks for models of short-term and working
  memory. \emph{Psychological Bulletin, 144}, 885-958.

}
\examples{
  data(bhatarah09b)
  ?bhatarah09c
}
\keyword{datasets}
